change SET

+++ controller/handlers_all.go
@@ -28,13 +28,13 @@ func NewAllCommandsImplementation(service service.AllCommandsService) AllCommand
 
 func (s allCommandsImplementation) SetAll(w http.ResponseWriter, r *http.Request) {
 
-	if remain.CheckMethod("PUT", r) {
+	if remain.CheckMethod("POST", r) {
 
 		var req service.RequestAll
 		err1 := json.NewDecoder(r.Body).Decode(&req)
 
 		if err1 != nil {
-			panic(err1)
+			w.WriteHeader(400)
 		}
 
 		resp := s.service.Set(req)
@@ -53,12 +53,12 @@ func (s *allCommandsImplementation) SaveAll(w http.ResponseWriter, r *http.Reque
 }
 
 func (s *allCommandsImplementation) AppendAll(w http.ResponseWriter, r *http.Request) {
-	if remain.CheckMethod("POST", r) {
+	if remain.CheckMethod("PUT", r) {
 		var req service.RequestAll
 		err1 := json.NewDecoder(r.Body).Decode(&req)
 
 		if err1 != nil {
-			panic(err1)
+			w.WriteHeader(400)
 		}
 		resp := s.service.Append(req)
 		fmt.Fprintf(w, "Resp: %+v", resp)
@@ -70,7 +70,15 @@ func (s *allCommandsImplementation) AppendAll(w http.ResponseWriter, r *http.Req
 func (s *allCommandsImplementation) GetAll(w http.ResponseWriter, r *http.Request) {
 	if remain.CheckMethod("GET", r) {
 		str := strings.Split(fmt.Sprint(r.URL), "B")
+		if len(str) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 		str1 := strings.Split(str[1], "%")
+		if len(str1) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 
 		var res *service.Ttl = s.service.Get(str1[0])
 		if res.ValueS != "nil" {
@@ -87,7 +95,15 @@ func (s *allCommandsImplementation) GetAll(w http.ResponseWriter, r *http.Reques
 func (s *allCommandsImplementation) DeleteKey(w http.ResponseWriter, r *http.Request) {
 	if remain.CheckMethod("DELETE", r) {
 		str := strings.Split(fmt.Sprint(r.URL), "B")
+		if len(str) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 		str1 := strings.Split(str[1], "%")
+		if len(str1) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 		resp := s.service.Delete(str1[0])
 
 		fmt.Fprintf(w, "Resp: %+v", resp)
