+++ controller/handlers_string.go
@@ -6,11 +6,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 	"strings"
 )
-
-
 
 type StringCommandsImplementation interface {
 	GetString(w http.ResponseWriter, r *http.Request)
@@ -20,97 +17,95 @@ type StringCommandsImplementation interface {
 	SaveString(w http.ResponseWriter, r *http.Request)
 }
 
-type RequestString struct{
-	Key string
+type RequestString struct {
+	Key   string
 	Value string
-	Ttl int64
+	Ttl   int64
 }
 
 type stringCommandsImplementation struct {
 	service service.StringCommandsService
 }
 
-func NewStringCommandsImplementation(service service.StringCommandsService) StringCommandsImplementation{
+func NewStringCommandsImplementation(service service.StringCommandsService) StringCommandsImplementation {
 	return &stringCommandsImplementation{service: service}
 }
 
+func (s *stringCommandsImplementation) SetString(w http.ResponseWriter, r *http.Request) {
 
-
-
-
-func(s *stringCommandsImplementation) SetString(w http.ResponseWriter, r *http.Request){
-
-
-	if remain.CheckMethod("PUT",r){
+	if remain.CheckMethod("POST", r) {
 		var req RequestString
 		err1 := json.NewDecoder(r.Body).Decode(&req)
 		if err1 != nil {
-			panic(err1)
+			w.WriteHeader(400)
 		}
 
-		s.service.Set(req.Key, req.Value, req.Ttl)
-	} else{
+		resp := s.service.Set(req.Key, req.Value, req.Ttl)
+		fmt.Fprintf(w, "Resp: %+v", resp)
+	} else {
 		w.WriteHeader(405)
 	}
 
-
-
-
 }
 
-func(s *stringCommandsImplementation) SaveString(w http.ResponseWriter, r *http.Request){
+func (s *stringCommandsImplementation) SaveString(w http.ResponseWriter, r *http.Request) {
 
 	s.service.Save()
 
 }
 
-
-
-
-
-
-
-func(s *stringCommandsImplementation) GetString(w http.ResponseWriter, r *http.Request){
-	if remain.CheckMethod("GET",r){
-		fmt.Println(url.Parse(fmt.Sprint(r.URL)))
+func (s *stringCommandsImplementation) GetString(w http.ResponseWriter, r *http.Request) {
+	if remain.CheckMethod("GET", r) {
 		str := strings.Split(fmt.Sprint(r.URL), "B")
+		if len(str) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 		str1 := strings.Split(str[1], "%")
+		if len(str1) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 
 		res := s.service.Get(str1[0])
-		if (len(res)!= 0){
+		if len(res) != 0 {
 			fmt.Fprintf(w, "Resp: %+v", res)
 		} else {
 			fmt.Fprintf(w, "Not Exist")
 		}
-	} else{
+	} else {
 		w.WriteHeader(405)
 	}
 
-
 }
 
-
-func(s *stringCommandsImplementation) DeleteKey(w http.ResponseWriter, r *http.Request){
-	if remain.CheckMethod("DELETE",r){
-		fmt.Println(url.Parse(fmt.Sprint(r.URL)))
+func (s *stringCommandsImplementation) DeleteKey(w http.ResponseWriter, r *http.Request) {
+	if remain.CheckMethod("DELETE", r) {
 		str := strings.Split(fmt.Sprint(r.URL), "B")
+		if len(str) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 		str1 := strings.Split(str[1], "%")
+		if len(str1) != 2 {
+			w.WriteHeader(400)
+			return
+		}
 
 		resp := s.service.Delete(str1[0])
 
 		fmt.Fprintf(w, "Resp: %+v", resp)
-	} else{
+	} else {
 		w.WriteHeader(405)
 	}
 }
 
-
-func(s *stringCommandsImplementation) GetKeys(w http.ResponseWriter, r *http.Request){
-	if remain.CheckMethod("GET",r){
+func (s *stringCommandsImplementation) GetKeys(w http.ResponseWriter, r *http.Request) {
+	if remain.CheckMethod("GET", r) {
 		tmp := s.service.Keys()
 
 		fmt.Fprintf(w, "Resp: %+v", tmp)
-	} else{
+	} else {
 		w.WriteHeader(405)
 	}
 }
